feat: add support for img tag
the img() tag can be called through a slot and will attach the contents of the image as an
additional message when sending to OpenAI. the image source can be a local path or remote url. if
the image is not a jpeg/png, it will be converted to jpeg format. also added a new ParserContext
object which encapsulates (existing) variables, options, and (new) attachments.

fix #39