feat: remove "assistant" from text outputs
when in "output assistant only" mode, the role "assistant" is no longer added to text logs saved
from the cli. this results in cleaner outputs that don't need to be trimmed.

BREAKING CHANGE: if your workflow relies on finding/removing "assistant" at the beginning of outputs
it will need to be updated

fix #47